refactor: rename /articles to /search for clarity
