+++ src/routes/api/index.js
@@ -1,5 +1,5 @@
 import express from 'express';
-import articleRoutes from './articlesRoutes.js';
+import searchRoutes from './searchRoutes.js';
 import authRoutes from './authRoutes.js';
 import bookmarkRoutes from './bookmarkRoutes.js';
 import newsRoutes from './newsRoutes.js';
@@ -10,7 +10,7 @@ const router = express.Router();
 router.use('/auth', authRoutes);
 router.use(protectedRoutes);
 router.use('/news', newsRoutes);
-router.use('/articles', articleRoutes);
+router.use('/search', searchRoutes);
 router.use('/bookmarks', bookmarkRoutes);
 
 export default router;
